requirements.txt
# Document parsing
PyMuPDF>=1.24.0
python-pptx>=1.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
pytesseract>=0.3.10
Pillow>=10.0.0

# LLM & Embeddings
openai>=1.30.0

# Vector store
chromadb>=0.5.0

# Text processing
tiktoken>=0.7.0

# Web server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# CLI & Config
python-dotenv>=1.0.0
rich>=13.0.0
click>=8.1.0
